Extract ImmSet applicative operations
