+++ src/Phunkie/Types/ImmSet.php
@@ -11,18 +11,17 @@
 
 namespace Phunkie\Types;
 
-use BadMethodCallException;
 use Phunkie\Cats\Applicative;
 use Phunkie\Cats\Show;
 use function Phunkie\Functions\show\showValue;
 use function Phunkie\Functions\type\promote;
+use Phunkie\Ops\ImmSet\ImmSetApplicativeOps;
 use Phunkie\Ops\ImmSet\ImmSetFunctorOps;
 use Phunkie\Utils\Iterator;
-use TypeError;
 
 class ImmSet implements Kind, Applicative
 {
-    use Show, ImmSetFunctorOps;
+    use Show, ImmSetFunctorOps, ImmSetApplicativeOps;
     private $elements = [];
 
     public function __construct(...$elements)
@@ -112,40 +111,6 @@ class ImmSet implements Kind, Applicative
             }
         }
         return ImmSet(...$new);
-    }
-
-    public function apply(Kind $f): Kind
-    {
-        $apply = function () use ($f) {
-            $result = [];
-            foreach($this->toArray() as $a) {
-                foreach ($f->toArray() as $ff) {
-                    if (!is_callable($ff)) {
-                        throw new TypeError(sprintf("`apply` takes Set<callable>, Set<%s> given", gettype($ff)));
-                    }
-                    $result[] = call_user_func($ff, $a);
-                }
-            }
-            return ImmSet(...$result);
-        };
-
-        switch (true) {
-            case $f == None(): return None();
-            case $f instanceof ImmSet: return $apply();
-            case $f instanceof Function1 && is_callable($f->get()):
-                return $this->map($f->get());
-            default: throw new BadMethodCallException();
-        }
-    }
-
-    public function map2(Kind $fb, callable $f): Kind
-    {
-        return $this->apply($fb->map(function($b) use ($f) { return function($a) use ($f, $b) { return $f($a, $b);};}));
-    }
-
-    public function pure($a): Kind
-    {
-        return ImmSet($a);
     }
 
     public function isEmpty(): bool
